backend/requirements.txt
# Backend requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
databricks-sdk>=0.12.0
databricks-sql-connector>=3.0.0
databricks-vectorsearch>=0.22
mlflow>=2.8.0
python-multipart>=0.0.6
